# specs/openapi.yaml
openapi: 3.0.0
info:
  title: API de GrowUp
  description: Especificación de la API para la plataforma de e-learning GrowUp.
  version: 1.1.0
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de Desarrollo Local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CourseStatus:
      type: string
      enum: [Publicado, Borrador, En Revision]
      description: Estado de publicación del curso
    CourseLevel:
      type: string
      enum: [Principiante, Intermedio, Avanzado]
      description: Nivel de dificultad del curso
...
    Topic:
      type: object
      required:
        - title
        - duration
      properties:
        title:
          type: string
          description: Título del tema
        duration:
          type: integer
          description: Duración en minutos
    Syllabus:
      type: object
      required:
        - title
        - description
        - topics
        - order
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del módulo
        title:
          type: string
          description: Título del módulo
        description:
          type: string
          description: Descripción del módulo
        order:
          type: integer
          description: Orden del módulo en el syllabus
        topics:
          type: array
          items:
            $ref: "#/components/schemas/Topic"
          description: Lista de temas del módulo
    Instructor:
      type: object
      required:
        - id
        - name
        - role
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del instructor
        name:
          type: string
          description: Nombre completo del instructor
        role:
          type: string
          description: Rol o título del instructor
        avatarUrl:
          type: string
          format: uri
          description: URL del avatar del instructor
        bio:
          type: string
          description: Biografía del instructor
    Course:
      type: object
      required:
        - name
        - category
        - price
        - publicationStatus
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del curso
        name:
          type: string
          description: Nombre del curso
        description:
          type: string
          description: Descripción del curso
        imageUrl:
          type: string
          format: uri
          description: URL de la imagen del curso
        category:
          type: string
          description: Categoría del curso (ej. DISEÑO WEB, DESARROLLO, ARQUITECTURA, MARKETING)
        level:
          $ref: "#/components/schemas/CourseLevel"
        price:
          type: number
          format: double
          description: Precio del curso
        duration:
          type: number
          description: Duración total en horas
        startDate:
          type: string
          format: date-time
          description: Fecha de inicio del curso
        endDate:
          type: string
          format: date-time
          description: Fecha de finalización del curso
        publicationStatus:
          $ref: "#/components/schemas/CourseStatus"
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación del curso
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última actualización
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Fecha de eliminación (soft delete)
        instructorId:
          type: string
          format: uuid
          description: ID del instructor (para desacoplamiento de bounded contexts)
        instructor:
          $ref: "#/components/schemas/Instructor"
        syllabus:
          type: array
          items:
            $ref: "#/components/schemas/Syllabus"
          description: Temario del curso
        enrolledCount:
          type: integer
          description: Número de estudiantes inscritos
        version:
          type: integer
          format: int64
          description: Versión de control de concurrencia

    # --- Modelos del Estudiante ---
...
    Review:
      type: object
      required:
        - id
        - courseId
        - rating
        - comment
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único de la reseña
        courseId:
          type: string
          format: uuid
          description: ID del curso valorado
        courseName:
          type: string
          description: Nombre del curso valorado
        studentId:
          type: string
          format: uuid
          description: ID del estudiante que valora
        studentName:
          type: string
          description: Nombre del estudiante que valora
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Calificación (1-5)
        comment:
          type: string
          description: Comentario de la reseña
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación
        version:
          type: integer
          format: int64
          description: Versión de control de concurrencia

    # --- DTOs de Peticiones/Respuestas ---
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: Código de estado HTTP
        message:
          type: string
          description: Mensaje de error
        timestamp:
          type: string
          format: date-time
          description: Momento del error

  # security: ELIMINADO - Se define seguridad por endpoint (ver paths)
paths:
...
  /courses:
    get:
      summary: Listar cursos
      description: Obtener lista de cursos con filtros opcionales
      tags: [Cursos]
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
          description: Filtrar por categoría
        - in: query
          name: level
          required: false
          schema:
            $ref: "#/components/schemas/CourseLevel"
          description: Filtrar por nivel
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/CourseStatus"
          description: Filtrar por estado de publicación
      responses:
        "200":
          description: Lista de cursos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"

    post:
      summary: Crear curso
      description: Crear un nuevo curso (requiere rol TEACHER o ADMIN)
      tags: [Cursos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
      responses:
        "201":
          description: Curso creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "403":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{id}:
    get:
      summary: Obtener detalles del curso
      description: Obtener información completa de un curso
      tags: [Cursos]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del curso
      responses:
        "200":
          description: Detalles del curso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Curso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Actualizar curso
      description: Actualizar información de un curso
      tags: [Cursos]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del curso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
      responses:
        "200":
          description: Curso actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Curso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Eliminar curso
      description: Eliminar un curso (soft delete)
      tags: [Cursos]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del curso
      responses:
        "204":
          description: Curso eliminado exitosamente
        "404":
          description: Curso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{id}/syllabus:
    get:
      summary: Obtener temario del curso
      description: Obtener el temario completo del curso
      tags: [Cursos]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del curso
      responses:
        "200":
          description: Temario del curso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Syllabus"
        "404":
          description: Curso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /courses/{id}/reviews:
    post:
      summary: Valorar un curso
      description: Crear una nueva reseña para un curso en el que el estudiante está inscrito
      tags: [Estudiante]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del curso a valorar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review"
      responses:
        "201":
          description: Reseña creada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Datos inválidos o estudiante no inscrito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Curso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...